+++ src/Lucid/Relations/BelongsToMany.js
@@ -220,6 +220,7 @@ class BelongsToMany extends BaseRelation {
    * @private
    */
   _decorateQuery () {
+    this._validateRead()
     this._selectFields()
     this._makeJoinQuery()
     this.wherePivot(this.foreignKey, this.$primaryKeyValue)
@@ -670,87 +671,6 @@ class BelongsToMany extends BaseRelation {
   addWhereOn (context) {
     this._makeJoinQuery()
     context.on(`${this.$primaryTable}.${this.primaryKey}`, '=', `${this.$pivotTable}.${this.foreignKey}`)
-  }
-
-  /**
-   * Prepare query for an aggregate function
-   *
-   * @method _prepareAggregate
-   */
-  _prepareAggregate () {
-    this._validateRead()
-    this._makeJoinQuery()
-    this.wherePivot(this.foreignKey, this.$primaryKeyValue)
-  }
-
-  /**
-   * Returns count of rows.
-   *
-   * @method count
-   *
-   * @param  {String} expression
-   *
-   * @return {Array}
-   */
-  count (expression) {
-    this._prepareAggregate()
-    return this.relatedQuery.count(expression)
-  }
-
-  /**
-   * Returns count of rows with distinct expression.
-   *
-   * @method count
-   *
-   * @param  {String} expression
-   *
-   * @return {Array}
-   */
-  countDistinct (expression) {
-    this._prepareAggregate()
-    return this.relatedQuery.countDistinct(expression)
-  }
-
-  /**
-   * Returns avg for a given column
-   *
-   * @method avg
-   *
-   * @param  {String} column
-   *
-   * @return {Array}
-   */
-  avg (column) {
-    this._prepareAggregate()
-    return this.relatedQuery.avg(column)
-  }
-
-  /**
-   * Return min value for a column
-   *
-   * @method avg
-   *
-   * @param  {String} column
-   *
-   * @return {Array}
-   */
-  min (column) {
-    this._prepareAggregate()
-    return this.relatedQuery.min(column)
-  }
-
-  /**
-   * Return max value for a column
-   *
-   * @method avg
-   *
-   * @param  {String} column
-   *
-   * @return {Array}
-   */
-  max (column) {
-    this._prepareAggregate()
-    return this.relatedQuery.max(column)
   }
 
   /**